requirements.txt
pymodbus>=3.5.4
pyyaml>=6.0.1
pyserial>=3.5
loguru>=0.7.2
pyserial-asyncio